AutoHire/requirements.txt
pymongo
pandas
xlsxwriter
pdfplumber
python-docx
imaplib2
pypandoc
docx2pdf
python-dotenv
openpyxl
scikit-learn